pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webhookapp"
version = "0.1.0"
description = "Trading-strategy management web application (Flask backend)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Nik Wekwerth", email = "nik@wekwerth.net" }
]

# Direct runtime deps (keep in sync with requirements*.txt if those exist)
dependencies = [
  "Flask>=2.3",
  "Flask-SQLAlchemy>=3.0",
  "Flask-Security-Too>=5.3",
  "Flask-Mail>=0.9",
  "Flask-Limiter>=3.5",
  "Flask-APScheduler>=1.12",
  "SQLAlchemy>=2.0",
  "ccxt>=4.0",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "pytz>=2024.1"
]

[project.optional-dependencies]
# Extra packages useful only in a development / test context
 dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-flask>=1.3",
    "freezegun>=1.4"
 ]

[tool.setuptools]
# The source code lives in the top-level "app" package
packages = ["app"]
# Also ship standalone modules at repo root (e.g. config.py)
py-modules = ["config"]
